pyproject.toml
[tool.poetry]
name = "certbot-certificate-hosttech"
version = "0.1.0"
description = "This repository contains the code needed to validate a domain name at hosttech.eu."
authors = ["Mose Mueller <mosmuell@ethz.ch>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.0"
certbot = "^2.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^5.0.4"
flake8-use-fstring = "^1.4"
flake8-functions = "^0.0.7"
flake8-comprehensions = "^3.11.1"
flake8-pep585 = "^0.1.7"
flake8-pep604 = "^0.1.0"
flake8-eradicate = "^1.4.0"
# flake8-expression-complexity = "^0.0.11"
mypy = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black" 

[tool.mypy]
show_error_codes = 1
disallow_untyped_defs = 1
disallow_untyped_calls = 1
disallow_incomplete_defs = 1
check_untyped_defs = 1